requirements.txt
ansible-core==2.19.7
azure-identity==1.25.0
azure-search-documents==11.6.0
azure-ai-projects==1.0.0
azure-ai-agents==1.0.0
openai==2.8.0
pydantic==2.11.4
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
azure-storage-blob==12.23.1